move interfaces to utils folder
