+++ src/handlers/AbstractLoginHandler.ts
@@ -2,9 +2,9 @@ import base64url from "base64url";
 
 import { UX_MODE } from "../utils/enums";
 import { broadcastChannelOptions, getTimeout, randomId } from "../utils/helpers";
+import { CreateHandlerParams, ILoginHandler, LoginWindowResponse, PopupResponse, TorusVerifierResponse } from "../utils/interfaces";
 import log from "../utils/loglevel";
 import { PopupHandler } from "../utils/PopupHandler";
-import { CreateHandlerParams, ILoginHandler, LoginWindowResponse, PopupResponse, TorusVerifierResponse } from "./interfaces";
 
 abstract class AbstractLoginHandler implements ILoginHandler {
   public nonce: string = randomId();
