+++ src/index.ts
@@ -1,7 +1,7 @@
 export { default as createHandler } from "./handlers/HandlerFactory";
-export * from "./handlers/interfaces";
 export { default as CustomAuth } from "./login";
 export * from "./utils/enums";
 export * from "./utils/error";
 export * from "./utils/helpers";
+export * from "./utils/interfaces";
 export * from "./utils/PopupHandler";
